added sound and functioning music
added sound and functioning music for level 1 and 2. The audiomanager and edited player move script are simple enough to use - just drag and drop the audio source components onto them.

I'll be able to get all the sound into the game, but may need help programming the other audio-based assets.